requirements.txt
langchain
faiss-cpu
transformers
sentence-transformers
streamlit
python-dotenv
requests
beautifulsoup4
